api_dj_project_1/dj_project_1/app1/viewsMpios.py
from rest_framework import viewsets
from rest_framework.permissions import IsAuthenticated, AllowAny
from rest_framework.authentication import TokenAuthentication
from .serializers import MpiosSerializer
from.models import Mpios

class MpiosViews(viewsets.ModelViewSet):
    serializer_class = MpiosSerializer
    queryset = Mpios.objects.all()
    authentication_classes = [TokenAuthentication]

    def get_permissions(self):
        if self.action in ['list', 'retrieve']:
            self.permission_classes = [AllowAny]
        else:
            self.permission_classes = [IsAuthenticated]
        return super().get_permissions()