api_dj_project_1/dj_project_1/requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==5.0.6
django-cors-headers==4.4.0
djangorestframework==3.15.2
gunicorn==22.0.0
idna==3.7
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
psycopg2==2.9.9
requests==2.32.3
sqlparse==0.5.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
